requirements.txt
streamlit
requests
beautifulsoup4
openai
Pillow
python-dotenv
playwright
